Splitting the data
The dataset was divided into two dataset. The one named "Train_data.csv", which has 691 data, is the data that should be used in training the model. The one named "Verification_data.csv" is the data that used to verify the model, which should not be used in the modeling process. The verification data were chosen randomly from the original dataset.